get rid of old gallery readme since it was moved, move tests out of 'core' and into a new folder escrow
